+++ src/Context/PersistContext.jsx
@@ -1,5 +1,4 @@
-import {createContext,useState,useEffect} from "react";
-import {auth} from "./../firebase";
+import {createContext,useState} from "react";
 
 const PersistContext=createContext();
 
@@ -8,24 +7,12 @@ export const PersistProvider=({children})=>{
     if(!localStorage.getItem("language"))
         localStorage.setItem("language","en");
 
-    
-    const [user,setUser]=useState();
-
-    useEffect(()=>{
-        const unsubscribe=auth.onAuthStateChanged((user)=>{
-            setUser(user);
-        })
-        return unsubscribe;
-    },[]);
-
     const [language,setLanguage]=useState(localStorage.getItem("language"));
     const [writeResponseText,setWriteResponseText]=useState("");
     const [analyzeResponse,setAnalyzeResponse]=useState(undefined); //object
     const [imagineResponseUrl,setImagineResponseUrl]=useState("");
 
     return (<PersistContext.Provider value={{
-        user,
-        setUser,
         language,
         setLanguage,
         writeResponseText,
